ecology_core/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-bio>=0.5.7
dendropy>=4.5.0
biom-format>=2.1.0
tqdm>=4.60.0